Update release script to create Homebrew bottles - improvements

+++ builder/cli-all-release.sh
@@ -16,28 +16,31 @@ function generatePlatformArchive() {
 
     PLATFORM_ARCHIVE_NAME=choreo-cli-$CHOREO_CLI_VERSION-${OS_PLATFORM}-${PLATFORM_ARCHITECTURE}.tar.gz
 
-    tar czf "$PLATFORM_ARCHIVE_NAME" "$TEMP_PLATFORM_BUILD_DIRECTORY_NAME" > /dev/null 2>&1
+    tar czf "$PLATFORM_ARCHIVE_NAME" "$TEMP_PLATFORM_BUILD_DIRECTORY_NAME"
     rm -rf "$TEMP_PLATFORM_BUILD_DIRECTORY"
 
     popd || exit 1
 }
 
 function generateHomebrewBottles() {
+    pushd "$TEMP_BUILD_DIRECTORY" || exit 1
+
+    MACOS_ARCHIVE=choreo-cli-$CHOREO_CLI_VERSION-macosx-x64.tar.gz
+    [ ! -f "$MACOS_ARCHIVE" ] && exit 1
+    tar xzf "$MACOS_ARCHIVE"
+
     TEMP_BOTTLE_DIRECTORY_NAME=chor
     TEMP_BOTTLE_DIRECTORY=$TEMP_BUILD_DIRECTORY/$TEMP_BOTTLE_DIRECTORY_NAME
 
-    AGGREGATED_GO_LDFLAGS="$GO_LDFLAGS -X $PROJECT_MODULE/internal/pkg/build.buildPlatform=darwin/amd64"
+    mkdir -p "$TEMP_BOTTLE_DIRECTORY_NAME"/"$CHOREO_CLI_VERSION"/bin
 
-    GOOS=darwin GOARCH=amd64 go build -o "$TEMP_BOTTLE_DIRECTORY"/"$CHOREO_CLI_VERSION"/bin/chor \
-                            -ldflags "$AGGREGATED_GO_LDFLAGS" -x "$PROJECT_MODULE/$CHORE_CLI_SRC_ROOT"
+    mv "$TEMP_PLATFORM_BUILD_DIRECTORY"/bin/chor "$TEMP_BOTTLE_DIRECTORY"/"$CHOREO_CLI_VERSION"/bin/chor
+    mv "$TEMP_PLATFORM_BUILD_DIRECTORY"/LICENSE "$TEMP_BOTTLE_DIRECTORY"/LICENSE
+    rm -rf "$TEMP_PLATFORM_BUILD_DIRECTORY"
 
-    cp "$PROJECT_ROOT"/LICENSE "$TEMP_BOTTLE_DIRECTORY"/"$CHOREO_CLI_VERSION"
-
-    pushd "$TEMP_BUILD_DIRECTORY" || exit 1
-
-    tar czf chor-"$CHOREO_CLI_VERSION".high_sierra.bottle.tar.gz $TEMP_BOTTLE_DIRECTORY_NAME > /dev/null 2>&1
-    tar czf chor-"$CHOREO_CLI_VERSION".mojave.bottle.tar.gz $TEMP_BOTTLE_DIRECTORY_NAME > /dev/null 2>&1
-    tar czf chor-"$CHOREO_CLI_VERSION".catalina.bottle.tar.gz $TEMP_BOTTLE_DIRECTORY_NAME > /dev/null 2>&1
+    tar czf chor-"$CHOREO_CLI_VERSION".high_sierra.bottle.tar.gz $TEMP_BOTTLE_DIRECTORY_NAME
+    tar czf chor-"$CHOREO_CLI_VERSION".mojave.bottle.tar.gz $TEMP_BOTTLE_DIRECTORY_NAME
+    tar czf chor-"$CHOREO_CLI_VERSION".catalina.bottle.tar.gz $TEMP_BOTTLE_DIRECTORY_NAME
     rm -rf "$TEMP_BOTTLE_DIRECTORY"
 
     popd || exit 1
